Refactor; Add: RAG for text messages
